feat(parsing): Add parsing of v2 of the settings specification
Fixes #6

BREAKING CHANGE: Loading build settings will from now on result in
a new structure. This structure has a dict in the task property, which
contains lists for the different task types. Example yaml of the new
structure:

    tasks:
      setup:
        - npm install
      tests:
        - npm test

Frigg yaml files on the old format is still supported, however, the
output of the parsing will represent the new structure.